requirements.txt
gunicorn
itsdangerous
flask
gtts
numpy
pillow
pytesseract
textblob
boto3
aiofiles==0.6.0
click==7.1.2
h11==0.11.0
MarkupSafe==1.1.1
pydantic==1.7
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3